requirements.txt
# requirements.txt
google-generativeai
requests
pdfplumber
python-docx
arxiv
langgraph
fastapi[all]
uvicorn[standard]